refactor(tui): split rendering responsibilities
